Split release and pre-release to handle multiple systems
